B7623 - Middle Tier, resetting the AR dashboard would cause an error Backported to: 17.4.73.0;18.1.14.0 Reviewed by: No one. Following middle tier patters.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/opendental18.1@19705 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/DashboardARs.cs
@@ -46,6 +46,10 @@ namespace OpenDentBusiness{
 
 		///<summary>Dashboardar is safe to truncate because it gets refilled as needed and there are no FKeys to any other table.</summary>
 		public static void Truncate() {
+			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
+				Meth.GetVoid(MethodBase.GetCurrentMethod());
+				return;
+			}
 			string command = "TRUNCATE dashboardar";
 			Db.NonQ(command);
 			if(!string.IsNullOrEmpty(PrefC.ReportingServer.Server)) { //only attempt to insert into the reporting server if the reporting server is set up.
